Adds useForm custom hook
