+++ registracija/src/components/ValidateInputs.js
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+const ValidateInputs = (props) => {
+  let errors = [];
+
+  return errors;
+};
+
+export default ValidateInputs;
